ui bug fixed, update word, blurred quiz

+++ lib/presenter/screens/vocabulary/view/vocabulary_page.dart
@@ -2,14 +2,17 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:language_learning/data/endpoint/word/update_word_endpoint.dart';
 import 'package:language_learning/generic/base_state.dart';
 import 'package:language_learning/presenter/screens/home/cubit/home_cubit.dart';
 import 'package:language_learning/presenter/screens/vocabulary/cubit/vocabulary_cubit.dart';
 import 'package:language_learning/presenter/screens/vocabulary/provider/vocabulary_provider.dart';
 import 'package:language_learning/presenter/widgets/primary_text.dart';
+import 'package:language_learning/presenter/widgets/primary_text_form_field.dart';
 import 'package:language_learning/utils/colors/app_colors.dart';
 import 'package:language_learning/utils/routes/navigation.dart';
 import 'package:provider/provider.dart';
+import 'package:flutter_tts/flutter_tts.dart';
 
 class VocabularyPage extends StatelessWidget {
   const VocabularyPage({super.key});
@@ -142,12 +145,31 @@ class VocabularyWordsList extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final homeCubit = context.read<HomeCubit>();
+    final FlutterTts flutterTts = FlutterTts();
+
+    Future<void> _initializeTTS() async {
+      try {
+        await flutterTts.awaitSpeakCompletion(true);
+        await flutterTts.setLanguage("en-US");
+        await flutterTts.setPitch(1.0);
+        await flutterTts.setSpeechRate(0.5);
+      } catch (e) {
+        print("TTS Initialization Error: $e");
+      }
+    }
+
+    Future<void> _speak(String text) async {
+      try {
+        await _initializeTTS();
+        await flutterTts.speak(text);
+      } catch (e) {
+        print("Error in TTS: $e");
+      }
+    }
 
     return BlocBuilder<VocabularyCubit, BaseState>(
       builder: (context, state) {
-        if (state is LoadingState) {
-          return const Center(child: CircularProgressIndicator());
-        } else if (state is SuccessState) {
+        if (state is SuccessState) {
           final data = state.data;
           return ListView.builder(
             itemCount: data.items.length,
@@ -156,10 +178,7 @@ class VocabularyWordsList extends StatelessWidget {
               final word = data.items[index];
 
               return Padding(
-                padding: EdgeInsets.symmetric(
-                  horizontal: 0.w,
-                  vertical: 5.h,
-                ),
+                padding: EdgeInsets.symmetric(horizontal: 0.w, vertical: 5.h),
                 child: Dismissible(
                   key: UniqueKey(),
                   background: Container(
@@ -180,33 +199,70 @@ class VocabularyWordsList extends StatelessWidget {
                       ],
                     ),
                   ),
-                  onDismissed: (direction) async {
+                  secondaryBackground: Container(
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(26.r),
+                      color: AppColors.primaryText.withValues(alpha: 0.5),
+                    ),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      children: [
+                        Padding(
+                          padding: EdgeInsets.only(right: 16.w),
+                          child: Icon(
+                            CupertinoIcons.pencil,
+                            color: AppColors.itemBackground,
+                            size: 20.w,
+                          ),
+                        )
+                      ],
+                    ),
+                  ),
+                  confirmDismiss: (direction) async {
                     if (direction == DismissDirection.startToEnd) {
-                      await vocabularyCubit.deleteWord(word.id);
-                      homeCubit.getLastWords();
-                      homeCubit.getCardCounts();
+                      vocabularyCubit.deleteWord(word.id);
+                      return true;
                     }
+                    if (direction == DismissDirection.endToStart) {
+                      _showUpdateDialog(context, vocabularyCubit, word);
+                      return false;
+                    }
+                    return false;
                   },
                   child: ListTile(
-                    contentPadding: EdgeInsets.symmetric(horizontal: 12.w, vertical: 2.h,),
+                    contentPadding:
+                        EdgeInsets.symmetric(horizontal: 12.w, vertical: 2.h),
                     tileColor: AppColors.unselectedItemBackground,
-                    shape: OutlineInputBorder(
-                      borderRadius: BorderRadius.circular(24).r,
-                      borderSide: BorderSide(color: Colors.transparent),
+                    shape: RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(24.r),
+                      side: BorderSide(color: Colors.transparent),
                     ),
-                    trailing: IconButton(
-                      onPressed: () async {
-                        await vocabularyCubit.addToLearning(word.id);
-                        homeCubit.getLastWords();
-                        homeCubit.getCardCounts();
-                        },
-                      icon: Icon(
-                        word.isLearningNow
-                            ? Icons.bookmark
-                            : Icons.bookmark_outline,
-                        size: 20.w,
-                        color: AppColors.bookMarkBackground,
-                      ),
+                    trailing: Row(
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        IconButton(
+                          onPressed: () => _speak(word.translation),
+                          icon: Icon(
+                            CupertinoIcons.volume_up,
+                            size: 20.w,
+                            color: AppColors.bookMarkBackground,
+                          ),
+                        ),
+                        IconButton(
+                          onPressed: () async {
+                            await vocabularyCubit.addToLearning(word.id);
+                            homeCubit.getLastWords();
+                            homeCubit.getCardCounts();
+                          },
+                          icon: Icon(
+                            word.isLearningNow
+                                ? Icons.bookmark
+                                : Icons.bookmark_outline,
+                            size: 20.w,
+                            color: AppColors.bookMarkBackground,
+                          ),
+                        ),
+                      ],
                     ),
                     title: PrimaryText(
                       text: '${word.source} - ${word.translation}',
@@ -220,8 +276,71 @@ class VocabularyWordsList extends StatelessWidget {
             },
           );
         } else {
-          return Center(child: Text('Loading...'));
+          return Center(child: Text('yuklenir'));
         }
+      },
+    );
+  }
+
+  void _showUpdateDialog(
+      BuildContext context, VocabularyCubit cubit, dynamic word) {
+    final TextEditingController sourceController =
+        TextEditingController(text: word.source);
+    final TextEditingController translationController =
+        TextEditingController(text: word.translation);
+
+    showDialog(
+      context: context,
+      builder: (context) {
+        return AlertDialog(
+          title: PrimaryText(
+            text: "Update Word",
+            color: AppColors.primaryText,
+            fontWeight: FontWeight.w400,
+            fontSize: 20,
+            fontFamily: 'DMSerifDisplay',
+          ),
+          content: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              PrimaryTextFormField(
+                onChanged: (value) {
+                  sourceController.text = value;
+                },
+                controller: sourceController,
+                headText: 'Source word',
+              ),
+              PrimaryTextFormField(
+                onChanged: (value) {
+                  translationController.text = value;
+                },
+                controller: translationController,
+                headText: 'Translation',
+              ),
+            ],
+          ),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.pop(context),
+              child: Text('Cancel'),
+            ),
+            TextButton(
+              onPressed: () async {
+                final homeCubit = context.read<HomeCubit>();
+                await cubit.updateWord(
+                  UpdateWordInput(
+                    id: word.id,
+                    source: sourceController.text,
+                    translation: translationController.text,
+                  ),
+                );
+                homeCubit.getLastWords();
+                Navigator.pop(context);
+              },
+              child: Text('Update'),
+            ),
+          ],
+        );
       },
     );
   }
